DEV: Упростим шаблоны

+++ app/Models/FeedModel.php
@@ -9,7 +9,7 @@ class FeedModel extends \Hleb\Scheme\App\Models\MainModel
 {
     public static function feed($page, $limit, $sheet, $slug)
     {
-        $user_id    = (int)UserData::getUserId();
+        $user_id    = UserData::getUserId();
         $string     = self::sorting($sheet);
 
         // Deleted post, banned from showing in the feed and limited by TL (trust_level)
@@ -86,7 +86,7 @@ class FeedModel extends \Hleb\Scheme\App\Models\MainModel
     public static function feedCount($sheet, $slug)
     {
         $string     = self::sorting($sheet);
-        $user_id    = (int)UserData::getUserId();
+        $user_id    = UserData::getUserId();
 
         $trust_level = ($user_id == 0) ? "AND post_tl = 0" : "AND post_tl <= " . UserData::getUserTl();
         $display = "AND post_is_deleted = 0 $trust_level";
